+++ lib/src/executable.dart
@@ -62,6 +62,16 @@ main(List<String> args) async {
       try {
         await _compileStylesheet(options, graph, source, destination);
       } on SassException catch (error, stackTrace) {
+        // This is an immediately-invoked function expression to work around
+        // dart-lang/sdk#33400.
+        () {
+          try {
+            if (destination != null) deleteFile(destination);
+          } on FileSystemException {
+            // If the file doesn't exist, that's fine.
+          }
+        }();
+
         printError(error.toString(color: options.color),
             options.trace ? stackTrace : null);