+++ lib/src/io/vm.dart
@@ -44,6 +44,8 @@ String readFile(String path) {
 void writeFile(String path, String contents) =>
     new io.File(path).writeAsStringSync(contents);
 
+void deleteFile(String path) => new io.File(path).deleteSync();
+
 Future<String> readStdin() async {
   var completer = new Completer<String>();
   completer.complete(await io.SYSTEM_ENCODING.decodeStream(io.stdin));